Run initdb automatically from the master leader
Summary:
Run initdb automatically from the master leader. On successful completion of initdb, persist that in
a boolean flag in SysYSQLCatalogConfigEntryPB. Also, if initdb has already started running, as
indicated by the existence of the pg_proc table with a fixed id, don't run it automatically.  This
is necessary for upgrading existing clusters. All of this behavior is currently under a flag, until
yb-ctl and all other YugaByte DB orchestration tools, as well as unit tests, are updated to avoid
running initdb. C++ based YSQL tests are being switched to the new behavior, but Java tests
will be switched later.

Test Plan:
- Jenkins

Manual test 1:
- `yb-ctl destroy`
- `yb-ctl create` (which will run initdb)
- `yb-ctl stop`
- `yb-ctl start --master_flags master_auto_run_initdb=true`
- Verify that initdb is not being run for the second time by the master. A line saying "Table pg_proc exists, assuming initdb has already been run" should appear in the master log.

Manual test 2:
- `yb-ctl destroy`
- `yb-ctl create --master_flags master_auto_run_initdb=true`
- Look at master logs and verify that it has run initdb.
- Try to connect with psql.

Reviewers: robert, neha, mihnea

Reviewed By: neha, mihnea

Subscribers: oleg, bogdan, yql, bharat

Differential Revision: https://phabricator.dev.yugabyte.com/D6501

+++ src/yb/yql/pgwrapper/pg_wrapper.cc
@@ -175,13 +175,30 @@ Status PgWrapper::InitDbForYSQL(const string& master_addresses, const string& tm
   std::mt19937 rng{std::random_device()()};
   conf.data_dir = Format("$0/tmp_pg_data_$1", tmp_dir_base, rng());
   BOOST_SCOPE_EXIT(&conf) {
-    Status del_status = Env::Default()->DeleteRecursively(conf.data_dir);
-    if (!del_status.ok()) {
-      LOG(WARNING) << "Failed to delete directory " << conf.data_dir;
+    auto is_dir = Env::Default()->IsDirectory(conf.data_dir);
+    if (is_dir.ok()) {
+      if (is_dir.get()) {
+        Status del_status = Env::Default()->DeleteRecursively(conf.data_dir);
+        if (!del_status.ok()) {
+          LOG(WARNING) << "Failed to delete directory " << conf.data_dir;
+        }
+      }
+    } else {
+      LOG(INFO) << "Failed to check directory existence for " << conf.data_dir << ": "
+                << is_dir.status();
     }
   } BOOST_SCOPE_EXIT_END;
   PgWrapper pg_wrapper(conf);
-  return pg_wrapper.InitDb(/* yb_enabled */ true);
+  auto start_time = std::chrono::steady_clock::now();
+  Status initdb_status = pg_wrapper.InitDb(/* yb_enabled */ true);
+  auto elapsed_time = std::chrono::steady_clock::now() - start_time;
+  LOG(INFO)
+      << "initdb took "
+      << std::chrono::duration_cast<std::chrono::milliseconds>(elapsed_time).count() << " ms";
+  if (!initdb_status.ok()) {
+    LOG(ERROR) << "initdb failed: " << initdb_status;
+  }
+  return initdb_status;
 }
 
 string PgWrapper::GetPostgresExecutablePath() {
